--- blocks/v2-country-check/v2-country-check.css
@@ -1,12 +1,0 @@
-.v2-country-check-container {
-    position: absolute;
-    top: 0;
-    left: 0;
-    z-index: 2;
-    background-color: var(--c-primary-white);
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width: 100vw;
-    height: 100vh;
-}
